+++ docs/source/changes.md
@@ -17,6 +17,8 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 - {pull}`395` refactors all occurrences of pybaum to {mod}`_pytask.tree_util`.
 - {pull}`396` replaces pybaum with optree and adds paths to the name of
   {class}`pytask.PythonNode`'s allowing for better hashing.
+- {class}`397` adds support for {class}`typing.NamedTuple` in
+  `@pytask.mark.task(kwargs=...)`.
 
 ## 0.3.2 - 2023-06-07